% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_draw.R
\name{bt_draw}
\alias{bt_draw}
\title{Draw the map}
\usage{
bt_draw(bertin, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{bertin}{bertin}

\item{width}{width}

\item{height}{height}

\item{elementId}{elementId}
}
\description{
Draw the map
}
