% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_geolines}
\alias{bt_geolines}
\title{Geolines}
\usage{
bt_geolines(bertin, stroke, strokeWidth, strokeDasharray, strokeLinecap, ...)
}
\arguments{
\item{bertin}{map object}

\item{stroke}{stroke}

\item{strokeWidth}{stroke width}

\item{strokeDasharray}{stroke dasharray}

\item{strokeLinecap}{stroke linecap}

\item{...}{other param}
}
\value{
a map object
}
\description{
Geolines
}
\examples{
library(sf)
world <- st_read(system.file("geojson/world.geojson", package = "bertin"))
bt_param(projection = "Eckert3", clip = TRUE) |>
  bt_outline() |>
  bt_layer(geojson = world, fill = "white") |>
  bt_geolines() |>
  bt_draw()
}
