% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_graticule}
\alias{bt_graticule}
\title{Graticules}
\usage{
bt_graticule(bertin, strokeWidth, strokeDasharray, stroke, step, ...)
}
\arguments{
\item{bertin}{map object}

\item{strokeWidth}{str wid}

\item{strokeDasharray}{str dash arr}

\item{stroke}{str}

\item{step}{step}

\item{...}{other param}
}
\value{
map object
}
\description{
Graticules
}
\examples{
library(sf)
mtq <- st_read(system.file("geojson/mtq.geojson", package = "bertin"), quiet = TRUE)
bt_param(margin = 10, width = 500, projection = "Winkel3") |>
  bt_layer(geojson = mtq, fill = "red") |>
  bt_graticule(strokeWidth = 1.5, stroke  = "green", step  = 0.1)|>
  bt_draw()
}
