% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_label}
\alias{bt_label}
\title{Labels}
\usage{
bt_label(
  bertin,
  data,
  values,
  fill = "#474342",
  fontSize = 10,
  fontFamily = "Roboto",
  textDecoration = "none",
  fontWeight = "normal",
  fontStyle = "normal",
  opacity = 1,
  halo = FALSE,
  halo_style = c("white", 4, 0.5),
  display = TRUE
)
}
\arguments{
\item{display}{}
}
\value{
map object
}
\description{
Labels
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
bt_param(projection = "Mercator", clip = TRUE) |>
  bt_layer(data = world) |>
  bt_label(data = world, values = "name") |>
  bt_draw()
}
