% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_layer}
\alias{bt_layer}
\title{Layer}
\usage{
bt_layer(
  bertin,
  data,
  fill = "randomcolor",
  tooltip,
  strokeWidth = 0.5,
  strokeDasharray = NULL,
  stroke = "white",
  strokeLinecap = "round",
  strokeLinejoin = "round",
  fillOpacity = 1,
  strokeOpacity = 1,
  symbol = "circle",
  symbol_size = 5,
  symbol_shift = 0,
  symbol_iteration = 200,
  display = TRUE
)
}
\arguments{
\item{bertin}{map object}

\item{data}{sf object EPSG:4326}

\item{fill}{fill color}

\item{tooltip}{tooltip allows you to display a label when you fly over an object}

\item{strokeWidth}{stroke width}

\item{strokeDasharray}{stroke dash array}

\item{stroke}{stroke color}

\item{strokeLinecap}{appearance of the line cap}

\item{strokeLinejoin}{appearance of the intersection beetween two or several lines}

\item{fillOpacity}{fill opacity}

\item{strokeOpacity}{stroke opacity}

\item{symbol}{if it is a dot layer, the type of symbol.circlecrossdiamondsquarestartrianglewye}

\item{symbol_size}{if it is a dot layer, a number indicating the size of the symbol}

\item{symbol_shift}{if it is a dot layer, use a value > 0 to switch symbols and avoid overlay}

\item{symbol_iteration}{Number of iteration to shift symbols}

\item{display}{Boolean to allow to show or hide the layer. This parameter has no effect on the calculation of the extent}
}
\value{
a bertin list
}
\description{
Layer
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
bt_param(margin = 10, width = 500, projection = "Winkel3") |>
  bt_layer(data = world, fill = "red") |>
  bt_draw()
}
