% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_mushroom}
\alias{bt_mushroom}
\title{Mushroom}
\usage{
bt_mushroom(
  bertin,
  data,
  top_values,
  bottom_values,
  bottom_tooltip,
  top_tooltip,
  ...
)
}
\arguments{
\item{bertin}{map object}

\item{data}{sf object EPSG:4326}

\item{top_values}{top values}

\item{bottom_values}{bottom values}

\item{bottom_tooltip}{bottom tooltip}

\item{top_tooltip}{top tooltip}

\item{...}{other param}
}
\value{
a map object
}
\description{
Mushroom
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
africa <- subset(world, region == "Africa")
africa$pop_pct <- (africa$pop / sum(africa$pop, na.rm = TRUE)) * 100
africa$gdp_pct <- (africa$gdp / sum(africa$gdp, na.rm = TRUE)) * 100
bt_layer(data = africa, fill = "#808080") |>
  bt_mushroom(data = africa, top_values = "gdp_pct", bottom_values = "pop_pct") |>
  bt_draw()
}
