% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_outline}
\alias{bt_outline}
\title{Outline}
\usage{
bt_outline(bertin, fill, opacity, stroke, strokeWidth, step, ...)
}
\arguments{
\item{bertin}{map object}

\item{fill}{fill}

\item{opacity}{opacity}

\item{stroke}{str}

\item{strokeWidth}{str wid}

\item{step}{step}

\item{...}{other param}
}
\value{
a map object
}
\description{
Outline
}
\examples{
bt_outline(fill = "#ADD8F7", stroke = "#4269ad") |>
  bt_draw()
}
