% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_param}
\alias{bt_param}
\title{Params}
\usage{
bt_param(margin, width, projection, ...)
}
\arguments{
\item{margin}{margin}

\item{width}{width}

\item{projection}{proj}

\item{...}{other params}
}
\value{
A list of params
}
\description{
Params
}
\examples{
bt_param(margin = 10)
}
