% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_regularbubble}
\alias{bt_regularbubble}
\title{Regular Bubble}
\usage{
bt_regularbubble(bertin, data, step, values, tooltip, ...)
}
\arguments{
\item{bertin}{map obj}

\item{data}{sf object EPSG:4326}

\item{step}{values}

\item{values}{k}

\item{tooltip}{tooltip}

\item{...}{other param}
}
\value{
a map object
}
\description{
Regular Bubble
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
bt_layer(data = world, fill = "#808080") |>
  bt_regularbubble(data = world,
                   values = "pop", step = 10,
                   k = 10, tooltip = c("$name","$pop")) |>
  bt_draw()
}
