% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_regulargrid}
\alias{bt_regulargrid}
\title{Regular Grid}
\usage{
bt_regulargrid(bertin, data, step, values, fill, tooltip, ...)
}
\arguments{
\item{bertin}{map obj}

\item{data}{sf object EPSG:4326}

\item{step}{step}

\item{values}{values}

\item{fill}{fill}

\item{tooltip}{tooltip}

\item{...}{other param}
}
\value{
a map object
}
\description{
Regular Grid
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
bt_layer(data = world, fill = "#808080") |>
  bt_regulargrid(data = world, values = "pop",
                 step = 20,
                fill = list(
                   nbreaks = 6,
                   method = "quantile",
                   colors = "Blues"
                 ),
                 tooltip = "$value") |>
  bt_draw()
}
