% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_rhumbs}
\alias{bt_rhumbs}
\title{Rhumbs}
\usage{
bt_rhumbs(
  bertin,
  nb = 10,
  position,
  stroke = "#786d6c",
  strokeWidth = 2,
  strokeOpacity = 1,
  strokeDasharray = "none",
  display = TRUE
)
}
\arguments{
\item{bertin}{map object}

\item{nb}{number of lines}

\item{position}{position of the text. It can be an array with x,y coordinates. For example [100,200]. It can be also a string defining the position.'topleft', 'top', 'topright', 'left', 'middle',
'right', 'bottomleft', 'bottom', 'bottomright'}

\item{stroke}{stroke color}

\item{strokeWidth}{stroke width}

\item{strokeOpacity}{stroke opacity}

\item{strokeDasharray}{stroke dash array}

\item{display}{Boolean to allow to show or hide the layer. This parameter has no effect on the calculation of the extent}
}
\value{
a map object
}
\description{
Rhumbs
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
bt_param(projection = "Eckert3", clip = TRUE) |>
  bt_outline() |>
  bt_layer(data = world, fill = "white") |>
  bt_rhumbs(position = c(370, 370), nb = 25) |>
  bt_draw()
}
