% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_export.R
\name{bt_save_svg}
\alias{bt_save_svg}
\title{Export to SVG}
\usage{
bt_save_svg(bertin, file, background, delay)
}
\arguments{
\item{bertin}{map object}

\item{file}{file}

\item{background}{background}

\item{delay}{delay (in seconds) to wait before saving the SVG}
}
\value{
a svg file
}
\description{
Export to SVG
}
\examples{
library(bertin)
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
map <- bt_layer(data = world, fill = "#808080") |>
bt_draw()
# bt_save_svg(bertin = map, file = "test.svg", delay = 0.5)
}
