% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_shadow}
\alias{bt_shadow}
\title{Shadow}
\usage{
bt_shadow(bertin, geojson, dx, dy, stdDeviation, opacity, ...)
}
\arguments{
\item{bertin}{map object}

\item{geojson}{geojson}

\item{dx}{dx}

\item{dy}{dy}

\item{stdDeviation}{blur}

\item{opacity}{opacity}

\item{...}{other param}
}
\value{
a map object
}
\description{
Shadow
}
\examples{
library(sf)
world <- st_read(system.file("geojson/world.geojson", package = "bertin"))
bt_param(projection = "Eckert3", clip = TRUE) |>
  bt_outline() |>
  bt_shadow(geojson = world) |>
  bt_layer(geojson = world, fill = "white") |>
  bt_draw()
}
