% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_smooth}
\alias{bt_smooth}
\title{Smooth}
\usage{
bt_smooth(bertin, data, values, thresholds, bandwidth, colorcurve, ...)
}
\arguments{
\item{bertin}{map object}

\item{data}{sf object EPSG:4326}

\item{values}{values}

\item{thresholds}{number of classes (default: 100)}

\item{bandwidth}{bandwidth (the standard deviation) of the Gaussian kernel and returns the estimate. (default: 5)}

\item{colorcurve}{a value to curve the color interpolation (default: 2)}

\item{...}{other param}
}
\value{
a map object
}
\description{
Smooth
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
bt_layer(data = world, fill = "#808080") |>
  bt_smooth(data = world, values = "pop",
            thresholds = 50,
            bandwidth = 25,
            colorcurve = 1) |>
  bt_draw()
}
