% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_spikes}
\alias{bt_spikes}
\title{Spikes}
\usage{
bt_spikes(
  bertin,
  data,
  values,
  k = 50,
  w = 10,
  fill = "#a31d88",
  stroke = 0.7,
  strokeWidth = 0.7,
  fillOpacity = 0.3,
  tooltip,
  display = TRUE,
  leg_x,
  leg_y,
  leg_fill = "none",
  leg_stroke = "black",
  leg_strokeWidth = 0.8,
  leg_txtcol = "#363636",
  leg_title = var_data,
  leg_round,
  leg_fontSize = 14,
  leg_fontSize2 = 10
)
}
\arguments{
\item{display}{}

\item{leg_fontSize2}{}
}
\value{
a map object
}
\description{
Spikes
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
bt_layer(data = world, fill = "#808080") |>
  bt_spikes(data = world, values = "pop", k = 110, w = 6) |>
  bt_draw()
}
