% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_square}
\alias{bt_square}
\title{Square}
\usage{
bt_square(bertin, geojson, values, k, tooltip, ...)
}
\arguments{
\item{bertin}{map obj}

\item{geojson}{sf obj}

\item{values}{values}

\item{k}{k}

\item{tooltip}{tooltip}

\item{...}{other param}
}
\value{
a map object
}
\description{
Square
}
\examples{
library(sf)
world <- st_read(system.file("geojson/world.geojson", package = "bertin"))
stats <- read.csv(system.file("csv/data.csv", package = "bertin"))
data <- merge(world, stats,  by.x = "ISO3", by.y = "id")
bt_param(projection = "Eckert3", clip = TRUE) |>
  bt_outline() |>
  bt_layer(geojson = world, fill = "white") |>
  bt_square(geojson = data,values = "pop", k = 60,
            tooltip = c("$NAMEen", "$pop", "(inh.)")) |>
  bt_draw()
}
