% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_text}
\alias{bt_text}
\title{Texts}
\usage{
bt_text(
  bertin,
  text,
  position = c(100, 200),
  fontSize = 15,
  fontFamily = "Roboto",
  textDecoration = "none",
  fontWeight = "normal",
  fontStyle = "none",
  margin = 0,
  anchor = "start",
  fill = "#474342",
  stroke = "none",
  frame_fill = "none",
  frame_stroke = "none",
  frame_strokeWidth = 1,
  frame_opacity = 1,
  display = TRUE
)
}
\arguments{
\item{display}{}
}
\value{
map object
}
\description{
Texts
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
bt_param(projection = "Mercator", clip = TRUE) |>
  bt_layer(data = world) |>
  bt_text(text = "This is my text", position = "bottomright",
          fontSize = 20, frame_stroke = "red", margin = 4) |>
  bt_draw()
}
