% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_tissot}
\alias{bt_tissot}
\title{Tissot's indicatrix}
\usage{
bt_tissot(
  bertin,
  step = 20,
  fill = "#d91848",
  fillOpacity = 0.5,
  stroke = "white",
  strokeWidth = 1.5,
  strokeOpacity = 1,
  display = TRUE
)
}
\value{
map object
}
\description{
#' @eval my_params(c(
}
\details{
'bertin',
 'step',
 'fill',
 'fillOpacity',
 'stroke',
 'strokeWidth',
 'strokeOpacity',
 'display',))
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
bt_param(projection = "Robinson", clip = TRUE) |>
  bt_outline() |>
  bt_layer(data = world, fill = "white", fillOpacity = .35) |>
  bt_tissot(step = 20) |>
  bt_draw()
}
