% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_bubble}
\alias{bt_bubble}
\title{Bubble}
\usage{
bt_bubble(bertin, geojson, values, k, tooltip, ...)
}
\arguments{
\item{bertin}{map obj}

\item{geojson}{sf obj}

\item{values}{values}

\item{k}{k}

\item{tooltip}{tooltip}

\item{...}{other param}
}
\value{
a map object
}
\description{
Bubble
}
\examples{
library(sf)
mtq <- st_read(system.file("geojson/mtq.geojson", package = "bertin"), quiet = TRUE)
bt_bubble(geojson = mtq, values = "POP", k = 60, tooltip = "$LIBGEO") |>
  bt_layer(geojson = mtq, fill = "#808080") |>
  bt_draw()
}
