% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_bubble}
\alias{bt_bubble}
\title{Bubble}
\usage{
bt_bubble(bertin, data, values, k, tooltip, ...)
}
\arguments{
\item{bertin}{map obj}

\item{data}{sf object EPSG:4326}

\item{values}{values}

\item{k}{k}

\item{tooltip}{tooltip}

\item{...}{other param}
}
\value{
a map object
}
\description{
Bubble
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
bt_layer(data = world, fill = "#808080") |>
  bt_bubble(data = world, values = "pop", k = 60, tooltip = "$LIBGEO") |>
  bt_draw()
}
