% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_dotcartogram}
\alias{bt_dotcartogram}
\title{Dot cartogram}
\usage{
bt_dotcartogram(
  bertin,
  data,
  onedot,
  nbmax = 200,
  iteration = 200,
  values,
  radius = 4,
  span = 0.5,
  fill = randomcolor,
  stroke = "white",
  strokeWidth = 0.5,
  fillOpacity = 1,
  tooltip,
  display = TRUE,
  leg_x,
  leg_y,
  leg_fill = "none",
  leg_stroke = "black",
  leg_strokeWidth = 0.8,
  leg_txtcol = "#363636",
  leg_title = var_data,
  leg_round,
  leg_fontSize = 10
)
}
\arguments{
\item{display}{}

\item{leg_fontSize}{}
}
\value{
a map object
}
\description{
Dot cartogram
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
bt_layer(data = world, fill = "#808080") |>
  bt_dotcartogram(data = world,
                  onedot = 100000000000,
                  iteration = 200, values = "gdp") |>
  bt_draw()
}
