% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_dotcartogram}
\alias{bt_dotcartogram}
\title{Dot cartogram}
\usage{
bt_dotcartogram(
  bertin,
  data,
  onedot,
  iteration,
  values,
  radius,
  span,
  fill,
  tooltip,
  ...
)
}
\arguments{
\item{bertin}{map obj}

\item{data}{sf object EPSG:4326}

\item{onedot}{dot value}

\item{iteration}{an integer to define the number of iteration for the Dorling method}

\item{values}{values}

\item{radius}{radius}

\item{span}{spacing between dots}

\item{fill}{fill color}

\item{tooltip}{tooltip}

\item{...}{other param}
}
\value{
a map object
}
\description{
Dot cartogram
}
\examples{
library(sf)
world <- st_read(system.file("geojson/world.geojson", package = "bertin"))
data <- read.csv(system.file("csv/data.csv", package = "bertin"))
bt_layer(data = world, fill = "#808080") |>
  bt_dotcartogram(data = merge(world, data,  by.x = "ISO3", by.y = "id"),
                  onedot = 200000000000, iteration = 200, values = "gdp") |>
  bt_draw()
}
