% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_dotcartogram}
\alias{bt_dotcartogram}
\title{Dot cartogram}
\usage{
bt_dotcartogram(
  bertin,
  data,
  onedot,
  nbmax = 200,
  iteration = 200,
  values,
  radius = 4,
  span = 0.5,
  fill,
  stroke = "white",
  strokeWidth = 0.5,
  fillOpacity = 1,
  tooltip,
  display = TRUE,
  leg_x,
  leg_y,
  leg_fill = "none",
  leg_stroke = "black",
  leg_strokeWidth = 0.8,
  leg_txtcol = "#363636",
  leg_title = values,
  leg_round,
  leg_fontSize = 10
)
}
\arguments{
\item{bertin}{map object}

\item{data}{sf object EPSG:4326}

\item{onedot}{dot value (if onedot is filled, nbmax is useless)}

\item{nbmax}{number max of circles on the map}

\item{iteration}{an integer to define the number of iteration for the Dorling method}

\item{values}{a string corresponding to the targeted variable in the properties}

\item{radius}{radius of dots}

\item{span}{spacing between dots}

\item{fill}{fill color}

\item{stroke}{stroke color}

\item{strokeWidth}{stroke width}

\item{fillOpacity}{fill opacity}

\item{tooltip}{tooltip allows you to display a label when you fly over an object}

\item{display}{Boolean to allow to show or hide the layer. This parameter has no effect on the calculation of the extent}

\item{leg_x}{position in x (if this value is not filled, the legend is not displayed)(value in px)}

\item{leg_y}{position in y (if this value is not filled, the legend is not displayed)(value in px)}

\item{leg_fill}{color of the circles}

\item{leg_stroke}{stroke color of the circles}

\item{leg_strokeWidth}{stoke-width of the legend}

\item{leg_txtcol}{color of the text}

\item{leg_title}{title of the legend}

\item{leg_round}{number of digits after the decimal point}

\item{leg_fontSize}{legend title font size}
}
\value{
a map object
}
\description{
Dot cartogram
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
bt_layer(data = world, fill = "#808080") |>
  bt_dotcartogram(data = world,
                  onedot = 100000000000,
                  iteration = 200, values = "gdp") |>
  bt_draw()
}
