% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_draw.R
\name{bt_draw}
\alias{bt_draw}
\title{Draw the map}
\usage{
bt_draw(bertin, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{bertin}{A `bertin` configuration list, typically initiated with [bt_param()].}

\item{width, height}{A numeric input in pixels.}

\item{elementId}{elementId}
}
\description{
Draw the map
}
