% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_geolines}
\alias{bt_geolines}
\title{Geolines}
\usage{
bt_geolines(
  bertin,
  stroke = "#020e21",
  strokeWidth = c(1.5, 1.2, 0.7),
  strokeDasharray = c("none", 5, 3),
  strokeLinecap = "round",
  display = TRUE
)
}
\arguments{
\item{bertin}{map object}

\item{stroke}{stroke color}

\item{strokeWidth}{stroke width}

\item{strokeDasharray}{stroke dash array}

\item{strokeLinecap}{appearance of the line cap}

\item{display}{Boolean to allow to show or hide the layer. This parameter has no effect on the calculation of the extent}

\item{strokeOpacity}{stroke opacity}
}
\value{
a map object
}
\description{
Geolines
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
bt_param(projection = "Eckert3", clip = TRUE) |>
  bt_outline() |>
  bt_layer(data = world, fill = "white") |>
  bt_geolines() |>
  bt_draw()
}
