% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_graticule}
\alias{bt_graticule}
\title{Graticules}
\usage{
bt_graticule(
  bertin,
  stroke = "white",
  strokeDasharray = "none",
  strokeWidth = 0.8,
  step = c(10, 10),
  display = TRUE
)
}
\arguments{
\item{bertin}{map object}

\item{stroke}{stroke color}

\item{strokeDasharray}{stroke dash array}

\item{strokeWidth}{stroke width}

\item{step}{gap between graticules. The value can be a number or an array of two values}

\item{display}{Boolean to allow to show or hide the layer. This parameter has no effect on the calculation of the extent}
}
\value{
map object
}
\description{
Graticules
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
bt_param(margin = 10, width = 500, projection = "Winkel3") |>
  bt_layer(data = world, fill = "red") |>
  bt_graticule(strokeWidth = 1.5, stroke  = "green", step  = 15)|>
  bt_draw()
}
