% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_graticule}
\alias{bt_graticule}
\title{Graticules}
\usage{
bt_graticule(
  bertin,
  stroke = "white",
  strokeDasharray = "none",
  strokeWidth = 0.8,
  step = c(10, 10),
  display = TRUE
)
}
\arguments{
\item{display}{}
}
\value{
map object
}
\description{
Graticules
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
bt_param(margin = 10, width = 500, projection = "Winkel3") |>
  bt_layer(data = world, fill = "red") |>
  bt_graticule(strokeWidth = 1.5, stroke  = "green", step  = 15)|>
  bt_draw()
}
