% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_hatch}
\alias{bt_hatch}
\title{Hatch}
\usage{
bt_hatch(
  bertin,
  angle = 45,
  stroke = "#786d6c",
  strokeWidth = 2,
  strokeOpacity = 1,
  strokeDasharray = "none",
  spacing = 8,
  display = TRUE
)
}
\arguments{
\item{display}{}
}
\value{
a map object
}
\description{
Hatch
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
bt_param(projection = "Eckert3", clip = TRUE) |>
  bt_outline() |>
  bt_layer(data = world, fill = "white") |>
  bt_hatch() |>
  bt_draw()
}
