% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_hatch}
\alias{bt_hatch}
\title{Hatch}
\usage{
bt_hatch(
  bertin,
  angle,
  stroke,
  strokeWidth,
  strokeOpacity,
  strokeDasharray,
  spacing,
  ...
)
}
\arguments{
\item{bertin}{map object}

\item{angle}{orientation of lines}

\item{stroke}{stroke}

\item{strokeWidth}{stroke width}

\item{strokeOpacity}{stroke opacity}

\item{strokeDasharray}{stroke dasharray}

\item{spacing}{spacing between lines}

\item{...}{other param}
}
\value{
a map object
}
\description{
Hatch
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
bt_param(projection = "Eckert3", clip = TRUE) |>
  bt_outline() |>
  bt_layer(data = world, fill = "white") |>
  bt_hatch() |>
  bt_draw()
}
