% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_header}
\alias{bt_header}
\title{Header}
\usage{
bt_header(
  bertin,
  text = "",
  fontSize = 20,
  fill = "#9e9696",
  background = "white",
  backgroundOpacity = 1,
  display = TRUE
)
}
\arguments{
\item{display}{}
}
\value{
a map object
}
\description{
Header
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
bt_param(margin = 10, width = 500, projection = "Winkel3") |>
  bt_layer(data = world, fill = "red") |>
  bt_header(text = "Title") |>
  bt_draw()
}
