% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_header}
\alias{bt_header}
\title{Header}
\usage{
bt_header(
  bertin,
  text = "",
  fontSize = 20,
  fill = "#9e9696",
  background = "white",
  backgroundOpacity = 1,
  display = TRUE
)
}
\arguments{
\item{bertin}{map object}

\item{text}{text to display}

\item{fontSize}{text size}

\item{fill}{fill color}

\item{background}{color of the background}

\item{backgroundOpacity}{background opacity}

\item{display}{Boolean to allow to show or hide the layer. This parameter has no effect on the calculation of the extent}

\item{anchor}{text anchor. start, middle, end}
}
\value{
a map object
}
\description{
Header
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
bt_param(margin = 10, width = 500, projection = "Winkel3") |>
  bt_layer(data = world, fill = "red") |>
  bt_header(text = "Title") |>
  bt_draw()
}
