% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_label}
\alias{bt_label}
\title{Labels}
\usage{
bt_label(
  bertin,
  data,
  values,
  fill = "#474342",
  fontSize = 10,
  fontFamily = "Roboto",
  textDecoration = "none",
  fontWeight = "normal",
  fontStyle = "normal",
  opacity = 1,
  halo = FALSE,
  halo_style = c("white", 4, 0.5),
  display = TRUE
)
}
\arguments{
\item{bertin}{map object}

\item{data}{sf object EPSG:4326}

\item{values}{a string corresponding to the targeted variable in the properties}

\item{fill}{fill color}

\item{fontSize}{text size}

\item{fontFamily}{font family}

\item{textDecoration}{text decoration.}

\item{fontWeight}{font weight. "normal", "bold", "bolder", "lighter"}

\item{fontStyle}{font style}

\item{opacity}{opacity of the object}

\item{halo}{boolean to display a halo below the text}

\item{halo_style}{an array defining the halo style. [color, thickness, opacity]}

\item{display}{Boolean to allow to show or hide the layer. This parameter has no effect on the calculation of the extent}
}
\value{
map object
}
\description{
Labels
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
bt_param(projection = "Mercator", clip = TRUE) |>
  bt_layer(data = world) |>
  bt_label(data = world, values = "name") |>
  bt_draw()
}
