% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_label}
\alias{bt_label}
\title{Labels}
\usage{
bt_label(bertin, data, values, fill, fontSize, fontFamily, ...)
}
\arguments{
\item{bertin}{map object}

\item{data}{sf object EPSG:4326}

\item{values}{values}

\item{fill}{fill}

\item{fontSize}{font size}

\item{fontFamily}{font family}

\item{...}{}
}
\value{
map object
}
\description{
Labels
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
bt_param(projection = "Mercator", clip = TRUE) |>
  bt_layer(data = world) |>
  bt_label(data = world, values = "name") |>
  bt_draw()
}
