% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_layer}
\alias{bt_layer}
\title{Layer}
\usage{
bt_layer(
  bertin,
  data,
  fill,
  tooltip,
  strokeWidth,
  strokeDasharray,
  stroke,
  step,
  ...
)
}
\arguments{
\item{bertin}{map object}

\item{data}{sf object EPSG:4326}

\item{fill}{fill}

\item{tooltip}{tooltip}

\item{strokeWidth}{stroke width}

\item{strokeDasharray}{stroke dash array}

\item{stroke}{stroke}

\item{step}{step}

\item{...}{other params}
}
\value{
a bertin list
}
\description{
Layer
}
\examples{
library(sf)
mtq <- st_read(system.file("geojson/mtq.geojson", package = "bertin"), quiet = TRUE)
bt_param(margin = 10, width = 500, projection = "Winkel3") |>
  bt_layer(data = mtq, fill = "red") |>
  bt_draw()
}
