% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_mushroom}
\alias{bt_mushroom}
\title{Mushroom}
\usage{
bt_mushroom(
  bertin,
  data,
  top_values,
  bottom_values,
  top_fill = "#d64f4f",
  bottom_fill = "#4fabd6",
  k = 50,
  stroke = "white",
  strokeWidth = 0.5,
  fillOpacity = 1,
  strokeOpacity = 1,
  top_tooltip,
  bottom_tooltip,
  display = TRUE,
  leg_x,
  leg_y,
  leg_fill = "none",
  leg_stroke = "black",
  leg_strokeWidth = 0.8,
  leg_txtcol = "#363636",
  leg_title,
  leg_round,
  leg_fontSize = 14,
  leg_fontSize2 = 10,
  leg_top_txt = top_values,
  leg_bottom_txt = bottom_values,
  leg_bottom_fill,
  leg_top_fill
)
}
\arguments{
\item{bertin}{map object}

\item{data}{sf object EPSG:4326}

\item{top_values}{a string corresponding to the targeted top variable in the properties (compulsory)}

\item{bottom_values}{a string corresponding to the targeted bottom variable in the properties(compulsory)}

\item{top_fill}{top fill color}

\item{bottom_fill}{bottom fill color}

\item{k}{size of the largest object}

\item{stroke}{stroke color}

\item{strokeWidth}{stroke width}

\item{fillOpacity}{fill opacity}

\item{strokeOpacity}{stroke opacity}

\item{top_tooltip}{top tooltip allows you to display a label when you fly over an object}

\item{bottom_tooltip}{tooltip allows you to display a label when you fly over an object}

\item{display}{Boolean to allow to show or hide the layer. This parameter has no effect on the calculation of the extent}

\item{leg_x}{position in x (if this value is not filled, the legend is not displayed)(value in px)}

\item{leg_y}{position in y (if this value is not filled, the legend is not displayed)(value in px)}

\item{leg_fill}{color of the circles}

\item{leg_stroke}{stroke color of the circles}

\item{leg_strokeWidth}{stoke-width of the legend}

\item{leg_txtcol}{color of the text}

\item{leg_title}{title of the legend}

\item{leg_round}{number of digits after the decimal point}

\item{leg_fontSize}{legend title font size}

\item{leg_fontSize2}{legend text font size}

\item{leg_top_txt}{title for the top variable}

\item{leg_bottom_txt}{title for the bottom variable}

\item{leg_bottom_fill}{color of bottom semi circles}

\item{leg_top_fill}{color of top semi circles}
}
\value{
a map object
}
\description{
Mushroom
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
africa <- subset(world, region == "Africa")
africa$pop_pct <- (africa$pop / sum(africa$pop, na.rm = TRUE)) * 100
africa$gdp_pct <- (africa$gdp / sum(africa$gdp, na.rm = TRUE)) * 100
bt_layer(data = africa, fill = "#808080") |>
  bt_mushroom(data = africa, top_values = "gdp_pct", bottom_values = "pop_pct") |>
  bt_draw()
}
