% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_mushroom}
\alias{bt_mushroom}
\title{Mushroom}
\usage{
bt_mushroom(
  bertin,
  data,
  top_values,
  bottom_values,
  top_fill = "#d64f4f",
  bottom_fill = "#4fabd6",
  k = 50,
  stroke = "white",
  strokeWidth = 0.5,
  fillOpacity = 1,
  strokeOpacity = 1,
  top_tooltip,
  bottom_tooltip,
  display = TRUE,
  leg_x,
  leg_y,
  leg_fill = "none",
  leg_stroke = "black",
  leg_strokeWidth = 0.8,
  leg_txtcol = "#363636",
  leg_title = var_data,
  leg_round,
  leg_fontSize = 14,
  leg_fontSize2 = 10,
  leg_top_txt = top_var,
  leg_bottom_txt = bottom_var,
  leg_bottom_fill,
  leg_top_fill
)
}
\arguments{
\item{display}{}

\item{leg_top_fill}{}
}
\value{
a map object
}
\description{
Mushroom
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
africa <- subset(world, region == "Africa")
africa$pop_pct <- (africa$pop / sum(africa$pop, na.rm = TRUE)) * 100
africa$gdp_pct <- (africa$gdp / sum(africa$gdp, na.rm = TRUE)) * 100
bt_layer(data = africa, fill = "#808080") |>
  bt_mushroom(data = africa, top_values = "gdp_pct", bottom_values = "pop_pct") |>
  bt_draw()
}
