% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_outline}
\alias{bt_outline}
\title{Outline}
\usage{
bt_outline(
  bertin,
  fill = "#add8f7",
  opacity = 1,
  stroke = "none",
  strokeWidth = 1,
  diplay = TRUE
)
}
\arguments{
\item{display}{}
}
\value{
a map object
}
\description{
Outline
}
\examples{
bt_outline(fill = "#ADD8F7", stroke = "#4269ad") |>
  bt_draw()
}
