% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_outline}
\alias{bt_outline}
\title{Outline}
\usage{
bt_outline(
  bertin,
  fill = "#add8f7",
  opacity = 1,
  stroke = "none",
  strokeWidth = 1,
  display = TRUE
)
}
\arguments{
\item{bertin}{map object}

\item{fill}{fill color}

\item{opacity}{opacity of the object}

\item{stroke}{stroke color}

\item{strokeWidth}{stroke width}

\item{display}{Boolean to allow to show or hide the layer. This parameter has no effect on the calculation of the extent}
}
\value{
a map object
}
\description{
Outline
}
\examples{
bt_outline(fill = "#ADD8F7", stroke = "#4269ad") |>
  bt_draw()
}
