% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_param}
\alias{bt_param}
\title{Params}
\usage{
bt_param(
  margin = 1,
  width = 1000,
  projection,
  extent = NULL,
  background = NULL,
  clip = FALSE,
  reverse = FALSE
)
}
\arguments{
\item{margin}{margin}

\item{width}{width in px}

\item{projection}{projection of the map}

\item{extent}{a feature or a bbox array defining the extent.It could be a country or [[112, -43],[153, -9]]}

\item{background}{color of the background}

\item{clip}{a boolean to avoid artifacts of discontinuous projection}

\item{reverse}{a boolean. By default, the layer placed on the topof the code is display on the top of the map.With reverse : true, your can reverse this order}
}
\value{
A list of params
}
\description{
Params
}
\examples{
bt_param(margin = 10)
}
