% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_regularbubble}
\alias{bt_regularbubble}
\title{Regular Bubble}
\usage{
bt_regularbubble(bertin, geojson, step, values, tooltip, ...)
}
\arguments{
\item{bertin}{map obj}

\item{geojson}{sf obj}

\item{step}{values}

\item{values}{k}

\item{tooltip}{tooltip}

\item{...}{other param}
}
\value{
a map object
}
\description{
Regular Bubble
}
\examples{
library(sf)
mtq <- st_read(system.file("geojson/mtq.geojson", package = "bertin"), quiet = TRUE)
bt_regularbubble(geojson = mtq, values = "POP", step = 30, k = 40, tooltip = c("$LIBGEO","$POP")) |>
  bt_layer(geojson = mtq, fill = "#808080") |>
  bt_draw()
}
