% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_regulargrid}
\alias{bt_regulargrid}
\title{Regular Grid}
\usage{
bt_regulargrid(
  bertin,
  data,
  step = 20,
  values,
  operator = "sum",
  geoprocessing = "intersection",
  blur = 0,
  colors = "Tableau10",
  order,
  col_missing = "#f5f5f5",
  txt_missing = "No data",
  stroke = "white",
  strokeWidth = 0.5,
  fillOpacity = 1,
  tooltip,
  leg_x,
  leg_y,
  leg_w = 30,
  leg_h = 20,
  leg_fill = "none",
  leg_stroke = "black",
  leg_strokeWidth = 0.8,
  leg_txtcol = "#363636",
  leg_title = var_data,
  leg_round,
  leg_fontSize = 10
)
}
\arguments{
\item{blur}{}

\item{tooltip}{}

\item{leg_fontSize}{}
}
\value{
a map object
}
\description{
Regular Grid
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
bt_layer(data = world, fill = "#808080") |>
  bt_regulargrid(data = world, values = "pop",
                 step = 20,
                fill = list(
                   nbreaks = 6,
                   method = "quantile",
                   colors = "Blues"
                 ),
                 tooltip = "$value") |>
  bt_draw()
}
