% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_regulargrid}
\alias{bt_regulargrid}
\title{Regular Grid}
\usage{
bt_regulargrid(
  bertin,
  data,
  step = 20,
  values,
  operator = "sum",
  geoprocessing = "intersection",
  blur = 0,
  colors = "Tableau10",
  order,
  col_missing = "#f5f5f5",
  txt_missing = "No data",
  stroke = "white",
  strokeWidth = 0.5,
  fillOpacity = 1,
  tooltip,
  leg_x,
  leg_y,
  leg_w = 30,
  leg_h = 20,
  leg_fill = "none",
  leg_stroke = "black",
  leg_strokeWidth = 0.8,
  leg_txtcol = "#363636",
  leg_title = values,
  leg_round,
  leg_fontSize = 10
)
}
\arguments{
\item{bertin}{map object}

\item{data}{sf object EPSG:4326}

\item{step}{gap between graticules. The value can be a number or an array of two values}

\item{values}{a string corresponding to the targeted variable in the properties}

\item{operator}{'sum' for absolute quantitative data or 'mean}

\item{geoprocessing}{'intersection' intersection between polygons and grid squares.
Assignment of values in proportion to the intersected area) or 'dotinpoly'
(considers each grid square and assigns the value of the underlying polygon
or the weighted average of intersecting polygons. ).}

\item{blur}{radius of the kernel defined}

\item{colors}{define the pallette of categorical colors}

\item{order}{an array of values to set the order of the colors}

\item{col_missing}{Color for missing values}

\item{txt_missing}{Text for missing values}

\item{stroke}{stroke color}

\item{strokeWidth}{stroke width}

\item{fillOpacity}{fill opacity}

\item{tooltip}{tooltip allows you to display a label when you fly over an object}

\item{leg_x}{position in x (if this value is not filled, the legend is not displayed)(value in px)}

\item{leg_y}{position in y (if this value is not filled, the legend is not displayed)(value in px)}

\item{leg_w}{width of the box}

\item{leg_h}{height of the box}

\item{leg_fill}{color of the circles}

\item{leg_stroke}{stroke color of the circles}

\item{leg_strokeWidth}{stoke-width of the legend}

\item{leg_txtcol}{color of the text}

\item{leg_title}{title of the legend}

\item{leg_round}{number of digits after the decimal point}

\item{leg_fontSize}{legend title font size}
}
\value{
a map object
}
\description{
Regular Grid
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
bt_layer(data = world, fill = "#808080") |>
  bt_regulargrid(data = world, values = "pop",
                 step = 20,
                fill = list(
                   nbreaks = 6,
                   method = "quantile",
                   colors = "Blues"
                 ),
                 tooltip = "$value") |>
  bt_draw()
}
