% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_regularsquare}
\alias{bt_regularsquare}
\title{Regular Square}
\usage{
bt_regularsquare(
  bertin,
  data,
  values,
  k = 50,
  fixmax,
  fill = randomcolor,
  stroke = "white",
  strokeWidth = 0.5,
  fillOpacity = 1,
  demrs = FALSE,
  iteration = 200,
  tooltip,
  display = TRUE,
  leg_x,
  leg_y,
  leg_fill = "none",
  leg_stroke = "black",
  leg_strokeWidth = 0.8,
  leg_txtcol = "#363636",
  leg_title = var_data,
  leg_round,
  leg_divisor = 1,
  leg_fontSize = 14,
  leg_fontSize2 = 10
)
}
\arguments{
\item{display}{}

\item{leg_fontSize2}{}
}
\value{
a map object
}
\description{
Regular Square
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
               layer = "world", quiet = TRUE)
bt_layer(data = world, fill = "#808080") |>
 bt_regularsquare(data = world,
                values = "pop", step = 10,
                  k = 10, tooltip = "$value") |>
 bt_draw()
}
