% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_rhumbs}
\alias{bt_rhumbs}
\title{Rhumbs}
\usage{
bt_rhumbs(
  bertin,
  nb = 10,
  position = c(height, 4, width, 4, width, 4),
  stroke = "#786d6c",
  strokeWidth = 2,
  strokeOpacity = 1,
  strokeDasharray = "none",
  display = TRUE
)
}
\value{
a map object
}
\description{
@eval my_params(c(
}
\details{
'bertin',
 'nb',
 'position',
 'stroke',
 'strokeWidth',
 'strokeOpacity',
 'strokeDasharray',
 'display'))
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
bt_param(projection = "Eckert3", clip = TRUE) |>
  bt_outline() |>
  bt_layer(data = world, fill = "white") |>
  bt_rhumbs(position = c(370, 370), nb = 25) |>
  bt_draw()
}
