% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_rhumbs}
\alias{bt_rhumbs}
\title{Rhumbs}
\usage{
bt_rhumbs(
  bertin,
  nb,
  position,
  stroke,
  strokeWidth,
  strokeOpacity,
  strokeDasharray,
  ...
)
}
\arguments{
\item{bertin}{map object}

\item{nb}{number of lines}

\item{position}{position of the center in x and y}

\item{stroke}{stroke color}

\item{strokeWidth}{stroke width}

\item{strokeOpacity}{stroke opacity}

\item{strokeDasharray}{stroke dasharray}

\item{...}{other param}
}
\value{
a map object
}
\description{
Rhumbs
}
\examples{
library(sf)
world <- st_read(system.file("geojson/world.geojson", package = "bertin"))
bt_param(projection = "Eckert3", clip = TRUE) |>
  bt_outline() |>
  bt_layer(geojson = world, fill = "white") |>
  bt_rhumbs(position = c(370, 370), nb = 25) |>
  bt_draw()
}
