% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_scalebar}
\alias{bt_scalebar}
\title{Scalebar}
\usage{
bt_scalebar(
  bertin,
  x = "left",
  y = "bottom",
  units = "kilometers",
  display = TRUE
)
}
\arguments{
\item{display}{}
}
\value{
a map object
}
\description{
Scalebar
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
bt_outline(fill = "#ADD8F7") |>
  bt_layer(data = world) |>
  bt_scalebar() |>
  bt_draw()
}
