% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_shadow}
\alias{bt_shadow}
\title{Shadow}
\usage{
bt_shadow(
  bertin,
  data,
  dx = 3,
  dy = 3,
  stdDeviation = 1.5,
  opacity = 0.7,
  fill = "#35383d",
  display = TRUE
)
}
\arguments{
\item{display}{}
}
\value{
a map object
}
\description{
Shadow
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
bt_param(projection = "Eckert3", clip = TRUE) |>
  bt_outline() |>
  bt_shadow(data = world) |>
  bt_layer(data = world, fill = "white") |>
  bt_draw()
}
