% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_smooth}
\alias{bt_smooth}
\title{Smooth}
\usage{
bt_smooth(
  bertin,
  data,
  values,
  stroke = "white",
  strokeWidth = 0.5,
  strokeLinecap = "round",
  strokeDasharray = "none",
  fllOpacity = 1,
  strokeOpacity = 1,
  display = TRUE,
  fill = "RdYlGn",
  thresholds = 100,
  bandwidth = 5,
  colorcurve = 2,
  reverse = FALSE,
  remove = 0,
  clip = FALSE,
  grid_step = 20,
  grid_blur = 0,
  grid_operator = "sum",
  grid_geoprocessing = "intersection"
)
}
\arguments{
\item{display}{}

\item{clip}{}
}
\value{
a map object
}
\description{
Smooth
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
bt_layer(data = world, fill = "#808080") |>
  bt_smooth(data = world, values = "pop",
            thresholds = 50,
            bandwidth = 25,
            colorcurve = 1) |>
  bt_draw()
}
