% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_spikes}
\alias{bt_spikes}
\title{Spikes}
\usage{
bt_spikes(
  bertin,
  data,
  values,
  k = 50,
  w = 10,
  fill = "#a31d88",
  stroke = 0.7,
  strokeWidth = 0.7,
  fillOpacity = 0.3,
  tooltip,
  display = TRUE,
  leg_x,
  leg_y,
  leg_fill = "none",
  leg_stroke = "black",
  leg_strokeWidth = 0.8,
  leg_txtcol = "#363636",
  leg_title = values,
  leg_round,
  leg_fontSize = 14,
  leg_fontSize2 = 10
)
}
\arguments{
\item{bertin}{map object}

\item{data}{sf object EPSG:4326}

\item{k}{size of the largest object}

\item{w}{width of the spikes}

\item{fill}{fill color}

\item{stroke}{stroke color}

\item{strokeWidth}{stroke width}

\item{fillOpacity}{fill opacity}

\item{tooltip}{tooltip allows you to display a label when you fly over an object}

\item{display}{Boolean to allow to show or hide the layer. This parameter has no effect on the calculation of the extent}

\item{leg_x}{position in x (if this value is not filled, the legend is not displayed)(value in px)}

\item{leg_y}{position in y (if this value is not filled, the legend is not displayed)(value in px)}

\item{leg_fill}{color of the circles}

\item{leg_stroke}{stroke color of the circles}

\item{leg_strokeWidth}{stoke-width of the legend}

\item{leg_txtcol}{color of the text}

\item{leg_title}{title of the legend}

\item{leg_round}{number of digits after the decimal point}

\item{leg_fontSize}{legend title font size}

\item{leg_fontSize2}{legend text font size}
}
\value{
a map object
}
\description{
Spikes
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
bt_layer(data = world, fill = "#808080") |>
  bt_spikes(data = world, values = "pop", k = 110, w = 6) |>
  bt_draw()
}
