% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_spikes}
\alias{bt_spikes}
\title{Spikes}
\usage{
bt_spikes(bertin, data, values, k, w, tooltip, ...)
}
\arguments{
\item{bertin}{map object}

\item{data}{sf object EPSG:4326}

\item{values}{values}

\item{k}{height of highest peak}

\item{w}{width of the spikes}

\item{tooltip}{tooltip}

\item{...}{other param}
}
\value{
a map object
}
\description{
Spikes
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
bt_layer(data = world, fill = "#808080") |>
  bt_spikes(data = world, values = "pop", k = 110, w = 6) |>
  bt_draw()
}
