% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_square}
\alias{bt_square}
\title{Square}
\usage{
bt_square(
  bertin,
  data,
  values,
  k = 50,
  fixmax,
  fill = "random color",
  stroke = "white",
  strokeWidth = 0.5,
  fillOpacity = 1,
  demers = FALSE,
  iteration = 200,
  tooltip,
  display = TRUE,
  leg_x,
  leg_y,
  leg_fill = "none",
  leg_stroke = "black",
  leg_strokeWidth = 0.8,
  leg_txtco = "#363636",
  leg_title = var_data,
  leg_round,
  leg_divisor = 1,
  leg_fontSize = 14,
  leg_fontSize2 = 10
)
}
\arguments{
\item{display}{}

\item{leg_fontSize2}{}
}
\value{
a map object
}
\description{
Square
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
               layer = "world", quiet = TRUE)
bt_layer(data = world, fill = "#808080") |>
bt_square(data = world, values = "pop", k = 60, tooltip = "$name") |>
bt_draw()
}
