% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_square}
\alias{bt_square}
\title{Square}
\usage{
bt_square(bertin, data, values, k, tooltip, ...)
}
\arguments{
\item{bertin}{map obj}

\item{data}{sf object EPSG:4326}

\item{values}{values}

\item{k}{k}

\item{tooltip}{tooltip}

\item{...}{other param}
}
\value{
a map object
}
\description{
Square
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
bt_param(projection = "Eckert3", clip = TRUE) |>
  bt_outline() |>
  bt_layer(data = world, fill = "white") |>
  bt_square(data = data,values = "pop", k = 60,
            tooltip = c("$name", "$pop", "(inh.)")) |>
  bt_draw()
}
