% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_text}
\alias{bt_text}
\title{Texts}
\usage{
bt_text(bertin, text, position, fontSize, fontFamily, ...)
}
\arguments{
\item{bertin}{map object}

\item{text}{text to display}

\item{position}{position}

\item{fontSize}{text size}

\item{fontFamily}{font family}

\item{...}{}
}
\value{
map object
}
\description{
Texts
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
bt_param(projection = "Mercator", clip = TRUE) |>
  bt_layer(data = world) |>
  bt_text(text = "This is my text", position = "bottomright",
          fontSize = 20, frame_stroke = "red", margin = 4) |>
  bt_draw()
}
