% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_text}
\alias{bt_text}
\title{Texts}
\usage{
bt_text(
  bertin,
  text,
  position = c(100, 200),
  fontSize = 15,
  fontFamily = "Roboto",
  textDecoration = "none",
  fontWeight = "normal",
  fontStyle = "none",
  margin = 0,
  anchor = "start",
  fill = "#474342",
  stroke = "none",
  frame_fill = "none",
  frame_stroke = "none",
  frame_strokeWidth = 1,
  frame_opacity = 1,
  display = TRUE
)
}
\arguments{
\item{bertin}{map object}

\item{text}{text to display}

\item{position}{position of the text. It can be an array with x,y coordinates. For example [100,200]. It can be also a string defining the position.'topleft', 'top', 'topright', 'left', 'middle',
'right', 'bottomleft', 'bottom', 'bottomright'}

\item{fontSize}{text size}

\item{fontFamily}{font family}

\item{textDecoration}{text decoration.}

\item{fontWeight}{font weight. "normal", "bold", "bolder", "lighter"}

\item{fontStyle}{font style}

\item{margin}{margin}

\item{anchor}{text anchor. start, middle, end}

\item{fill}{fill color}

\item{stroke}{stroke color}

\item{frame_fill}{frame background color}

\item{frame_stroke}{frame stroke color}

\item{frame_strokeWidth}{thickness of the frame contour}

\item{frame_opacity}{frame opacity}

\item{display}{Boolean to allow to show or hide the layer. This parameter has no effect on the calculation of the extent}
}
\value{
map object
}
\description{
Texts
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
bt_param(projection = "Mercator", clip = TRUE) |>
  bt_layer(data = world) |>
  bt_text(text = "This is my text", position = "bottomright",
          fontSize = 20, frame_stroke = "red", margin = 4) |>
  bt_draw()
}
