% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_tiles}
\alias{bt_tiles}
\title{Mercator Tiles}
\usage{
bt_tiles(bertin, style, zoomDelta, tileSize, opacity, ...)
}
\arguments{
\item{bertin}{map object}

\item{style}{tile style}

\item{zoomDelta}{zoom offset}

\item{tileSize}{tile size}

\item{opacity}{tile opacity}

\item{...}{}
}
\value{
map object
}
\description{
Mercator Tiles
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
italy <- world[world$name == "Italy",]
bt_param(projection = "Mercator", clip = TRUE, extent = italy) |>
  bt_tiles(style = "worldphysical") |>
  bt_draw()
}
