% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_tissot}
\alias{bt_tissot}
\title{Tissot's indicatrix}
\usage{
bt_tissot(
  bertin,
  step = 20,
  fill = "#d91848",
  fillOpacity = 0.5,
  stroke = "white",
  strokeWidth = 1.5,
  strokeOpacity = 1,
  display = TRUE
)
}
\arguments{
\item{bertin}{map object}

\item{step}{gap between graticules. The value can be a number or an array of two values}

\item{fill}{fill color}

\item{fillOpacity}{fill opacity}

\item{stroke}{stroke color}

\item{strokeWidth}{stroke width}

\item{strokeOpacity}{stroke opacity}

\item{display}{Boolean to allow to show or hide the layer. This parameter has no effect on the calculation of the extent}
}
\value{
map object
}
\description{
Tissot's indicatrix
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
bt_param(projection = "Robinson", clip = TRUE) |>
  bt_outline() |>
  bt_layer(data = world, fill = "white", fillOpacity = .35) |>
  bt_tissot(step = 20) |>
  bt_draw()
}
