% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_tissot}
\alias{bt_tissot}
\title{Tissot's indicatrix}
\usage{
bt_tissot(
  bertin,
  step,
  fill,
  fillOpacity,
  stroke,
  strokeWidth,
  strokeOpacity,
  ...
)
}
\arguments{
\item{bertin}{map object}

\item{step}{step between circles}

\item{fill}{fill}

\item{fillOpacity}{fill opacity}

\item{stroke}{stroke}

\item{strokeWidth}{stroke width}

\item{strokeOpacity}{stroke opacity}

\item{...}{other param}
}
\value{
map object
}
\description{
Tissot's indicatrix
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
bt_param(projection = "Robinson", clip = TRUE) |>
  bt_outline() |>
  bt_layer(data = world, fill = "white", fillOpacity = .35) |>
  bt_tissot(step = 20) |>
  bt_draw()
}
