% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_waterlines}
\alias{bt_waterlines}
\title{Waterlines}
\usage{
bt_waterlines(
  bertin,
  data,
  dist = 5,
  nb = 3,
  precision = 3,
  stroke = "#5d81ba",
  strokeOpacity = c(1, 0.1),
  strokeWidth = c(1.2, 0.2),
  strokeDasharray = "none",
  strokeLinecap = "round",
  strokeLinejoin = "round",
  display = TRUE
)
}
\arguments{
\item{bertin}{map object}

\item{data}{sf object EPSG:4326}

\item{dist}{distance between lines in pixels}

\item{nb}{number of lines}

\item{precision}{number of steps defining buffers. 1 = not accurate, 8 = accurate}

\item{stroke}{stroke color}

\item{strokeOpacity}{stroke opacity}

\item{strokeWidth}{stroke width}

\item{strokeDasharray}{stroke dash array}

\item{strokeLinecap}{appearance of the line cap}

\item{strokeLinejoin}{appearance of the intersection beetween two or several lines}

\item{display}{Boolean to allow to show or hide the layer. This parameter has no effect on the calculation of the extent}
}
\value{
a map object
}
\description{
Waterlines
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
bt_param(margin = 10, width = 500, projection = "NellHammer") |>
  bt_layer(data = world) |>
  bt_waterlines(data = world, dist = 3, nb = 4) |>
  bt_draw()
}
