% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_waterlines}
\alias{bt_waterlines}
\title{Waterlines}
\usage{
bt_waterlines(bertin, geojson, dist, nb, ...)
}
\arguments{
\item{bertin}{map object}

\item{geojson}{sf}

\item{dist}{dist}

\item{nb}{nb}

\item{...}{other}
}
\value{
a map object
}
\description{
Waterlines
}
\examples{
library(sf)
mtq <- st_read(system.file("geojson/mtq.geojson", package = "bertin"), quiet = TRUE)
bt_param(margin = 10, width = 500, projection = "Winkel3") |>
  bt_layer(geojson = mtq, fill = "red") |>
  bt_waterlines(geojson = mtq, dist = 0.25, nb = 5)|>
  bt_draw()
}
