% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_functions.R
\name{bt_waterlines}
\alias{bt_waterlines}
\title{Waterlines}
\usage{
bt_waterlines(bertin, data, dist, nb, ...)
}
\arguments{
\item{bertin}{map object}

\item{data}{sf object EPSG:4326}

\item{dist}{dist}

\item{nb}{nb}

\item{...}{other}
}
\value{
a map object
}
\description{
Waterlines
}
\examples{
library(sf)
world <- st_read(system.file("gpkg/world.gpkg", package = "bertin"),
                 layer = "world", quiet = TRUE)
bt_param(margin = 10, width = 500, projection = "NellHammer") |>
  bt_layer(data = world) |>
  bt_waterlines(data = world, dist = 3, nb = 4) |>
  bt_draw()
}
